Fix remaining sync response Debug event content leaks
